Flow control ability in Dispatcher
- documentation fixes
- tests moved to common file test_updater.py
- @property Promise.exception added
- log warning in case of using DispatcherHandlerFlow in executed method asynchronously
